requirements/requirements-dev.txt
black==23.1.0
django_debug_toolbar==3.8.1
django-tz-detect==0.4.0
isort==5.11.3
flake8==5.0.4
flake8-import-order==0.18.2
flake8-quotes==3.3.2
sorl-thumbnail==12.9.0